third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/show-function-definition-expected.txt

Tests that "Show Function Definition" jumps to the correct location.


Running: testRevealFunctionDefinition
Function location revealed: [7:18]

Running: testDumpFunctionDefinition
jumpToMe
function jumpToMe()
{
    var result = 12345;
    return window.foo || result;
}
